+++ src/handlers.rs
@@ -460,12 +460,14 @@ pub async fn set_keybind(
 }
 
 // TODO: Should be select
+// TODO: max should be dynamic
 /// Renders port input for MIDI port
 fn port_cell(uuid: &str, associated_port: usize) -> Markup {
     html! {
         input
             type="number" value=(format!("{}", associated_port))
             name="port"
+            min=(MIN_PORT_NUMBER)
             hx-target="this"
             hx-swap="outerHTML"
             hx-post=(format!("/blocks/midi/set-port/{uuid}"));
@@ -478,6 +480,12 @@ pub struct SetPort {
     /// MIDI port to set for Midi block
     pub port: usize,
 }
+
+/// Minimum MIDI port number
+///
+/// On windows it is 1 since we don't have virtual ports.
+/// On unix'es it's 0 since we have virtual ports.
+const MIN_PORT_NUMBER: usize = if cfg!(unix) { 0 } else { 1 };
 
 /// Set port for MIDI block
 pub async fn set_port_for_midi(
@@ -497,10 +505,11 @@ pub async fn set_port_for_midi(
         return Err(StatusCode::BAD_REQUEST);
     };
 
-    let min = 0_usize;
     let max = app_state.connection.read().unwrap().ports.len();
-    if port < min || port > max {
-        error!("port number should be between {min} and {max}");
+
+    #[allow(clippy::absurd_extreme_comparisons)]
+    if port < MIN_PORT_NUMBER || port > max {
+        error!("port number should be between {MIN_PORT_NUMBER} and {max}");
         return Ok(port_cell(&uuid, midi.associated_port));
     }